test: changed proxy logic
