+++ metro.config.js
@@ -108,6 +108,19 @@ module.exports = function (baseConfig) {
               type: 'sourceFile',
             };
           }
+          if (
+            !isE2E &&
+            (moduleName === './SnapsE2EProxy' ||
+              moduleName.endsWith('/SnapsE2EProxy'))
+          ) {
+            return {
+              type: 'sourceFile',
+              filePath: path.resolve(
+                __dirname,
+                'app/lib/snaps/SnapsE2EProxy.stub.ts',
+              ),
+            };
+          }
           if (isE2E) {
             if (moduleName === '@sentry/react-native') {
               return {
