+++ tests/smoke/confirmations/send/send-btc-token.spec.ts
@@ -6,11 +6,10 @@ import { withFixtures } from '../../../framework/fixtures/FixtureHelper';
 import FixtureBuilder from '../../../framework/fixtures/FixtureBuilder';
 import { loginToApp } from '../../../flows/wallet.flow';
 
-// Ported flow from metamask-extension/test/e2e/tests/send/btc-send.spec.ts
 const TOKEN = 'Bitcoin';
 
 describe(SmokeConfirmations('Send Bitcoin'), () => {
-  it('shows insufficient funds for a high BTC amount', async () => {
+  it('shows insufficient funds', async () => {
     await withFixtures(
       {
         fixture: new FixtureBuilder().build(),
@@ -22,7 +21,7 @@ describe(SmokeConfirmations('Send Bitcoin'), () => {
         await WalletView.waitForTokenToBeReady(TOKEN, 0);
         await WalletView.tapOnToken(TOKEN);
         await TokenOverview.tapSendButton();
-        await SendView.pressAmountFiveButton();
+        await SendView.enterZeroAmount();
         await SendView.checkInsufficientFundsError();
       },
     );
